Renamed Login and Chat components
Renamed to LoginPage and ChatPage for clarity
