+++ client/src/components/App.tsx
@@ -1,7 +1,7 @@
 import { Routes, Route, Navigate } from 'react-router-dom';
-import Login from './Login';
+import LoginPage from './LoginPage';
 import { Container } from '@mui/material';
-import Chat from './MainPage';
+import ChatPage from './ChatPage';
 
 import { ToastContainer } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
@@ -17,10 +17,10 @@ const App = () => {
     <>
       <Container>
         <Routes>
-          <Route path='/login' element={currentUser ?  <Navigate to='/' replace /> : <Login />} />
+          <Route path='/login' element={currentUser ?  <Navigate to='/' replace /> : <LoginPage />} />
           <Route
             path='/*'
-            element={currentUser ? <Chat /> : <Navigate to='/login' replace />}
+            element={currentUser ? <ChatPage /> : <Navigate to='/login' replace />}
           />
           <Route path='*' element={<Navigate to='/' replace />} />
         </Routes>
