+++ client/src/components/ChatPage/NavBar.tsx
